use proper RelayChainSlot type

+++ pallets/parachain-staking/src/tests.rs
@@ -6723,7 +6723,7 @@ fn deferred_payment_steady_state_event_flow() {
 						total_exposed_amount: 400,
 					},
 					Event::NewRound {
-						starting_block: (((round - 1) * 5) * 2) + 1,
+						starting_block: (((round as u64 - 1) * 5) * 2) + 1,
 						round: round as u32,
 						selected_collators_number: 4,
 						total_balance: 1600,
